# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "My API",
    "description": "Description",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/contacts/": {
      "get": {
        "description": "Gets all contacts in the database",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "Inserts a document into the database with the info you provide in json format",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "Jane"
                },
                "lastName": {
                  "example": "Doe"
                },
                "email": {
                  "example": "janedoe@mail.com"
                },
                "favoriteColor": {
                  "example": "white"
                },
                "birthday": {
                  "example": "0000-00-00"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/contacts/{id}": {
      "get": {
        "description": "Gets one contact that matches the id given by you. Test it out using this ID '627049cf47918eab3a1e82f0', or any id returned with GET /contacts/",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "description": "Updates one contact that matches the id given by you. Test it out by first scrolling up and using POST /contacts/ to create a contact, then copy the 'insertedId' returned in the response body and come back here to update it",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "firstName": {
                  "example": "Jane"
                },
                "lastName": {
                  "example": "Doe"
                },
                "email": {
                  "example": "updated email"
                },
                "favoriteColor": {
                  "example": "updated color"
                },
                "birthday": {
                  "example": "0000-00-00"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "Deletes one contact that matches the id given by you. Test it out by first scrolling up and using POST /contacts/ to create a contact, then copy the 'insertedId' returned in the response body and come back here to delete it",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  }
}